// backend/types/item.d.ts
// export interface Item{
//     brand: string
//     category: string
//     description: string
//     discountPercentage: number
//     id: number
//     images: string[]
//     price: number
//     rating: number
//     stock: number
//     thumbnail: string
//     title: string
// }

export interface DBItemInfo{
    _id: string,
    id: number,
    quantity: number,
    __v: number
  }